Moved page actions to it's own use case
